[ADF-968] - moved site dropdown into documentlist
