+++ demo-shell-ng2/app/vendor.ts
@@ -32,7 +32,6 @@ import 'ng2-alfresco-upload';
 import 'ng2-alfresco-userinfo';
 import 'ng2-alfresco-viewer';
 import 'ng2-alfresco-webscript';
-import 'ng2-alfresco-sites';
 
 // Polyfill(s) for dialogs
 require('script-loader!dialog-polyfill/dialog-polyfill');
