+++ ng2-components/ng2-alfresco-documentlist/src/models/site.model.ts
