+++ ng2-components/ng2-alfresco-documentlist/src/services/sites.service.spec.ts
