+++ ng2-components/ng2-alfresco-documentlist/src/services/sites.service.ts
