--- ng2-components/ng2-alfresco-sites/.editorconfig
@@ -1,23 +1,0 @@
-# http://editorconfig.org
-
-root = true
-
-[*]
-charset = utf-8
-indent_style = space
-indent_size = 4
-end_of_line = lf
-insert_final_newline = true
-trim_trailing_whitespace = true
-
-[package.json]
-indent_style = space
-indent_size = 2
-
-[karma.conf.js]
-indent_style = space
-indent_size = 2
-
-[*.md]
-insert_final_newline = false
-trim_trailing_whitespace = false
