--- ng2-components/ng2-alfresco-sites/.gitignore
@@ -1,20 +1,0 @@
-npm-debug.log
-node_modules
-.idea
-.npmrc
-typings
-coverage
-dist
-src/**/*.js
-src/**/*.js.map
-src/**/*.d.ts
-demo/**/*.js.map
-demo/**/*.d.ts
-index.js
-index.js.map
-!systemjs.config.js
-*.tgz
-/package/
-/bundles/
-index.d.ts
-/.happypack
