--- ng2-components/ng2-alfresco-sites/.npmignore
@@ -1,19 +1,0 @@
-npm-debug.log
-.idea
-.npmrc
-
-coverage/
-demo/
-dist/
-node_modules
-typings/
-fonts/
-
-/.editorconfig
-/.travis.yml
-/*.json
-/karma-test-shim.js
-/karma.conf.js
-/gulpfile.ts
-/.npmignore
-/.happypack
