--- ng2-components/ng2-alfresco-sites/README.md
@@ -1,90 +1,0 @@
-# Alfresco Sites Component
-
-<!-- markdown-toc start - Don't edit this section.  npm run toc to generate it-->
-
-<!-- toc -->
-
-- [Prerequisites](#prerequisites)
-- [Install](#install)
-- [Basic usage](#basic-usage)
-- [adf-sites-dropdown](#adf-sites-dropdown)
-  * [Properties](#properties)
-  * [Events](#events)
-- [Build from sources](#build-from-sources)
-- [NPM scripts](#npm-scripts)
-- [Demo](#demo)
-- [License](#license)
-
-<!-- tocstop -->
-
-<!-- markdown-toc end -->
-
-## Prerequisites
-
-Before you start using this development framework, make sure you have installed all required software and done all the
-necessary configuration [prerequisites](https://github.com/Alfresco/alfresco-ng2-components/blob/master/PREREQUISITES.md).
-
-> If you plan using this component with projects generated by Angular CLI, please refer to the following article: [Using ADF with Angular CLI](https://github.com/Alfresco/alfresco-ng2-components/wiki/Angular-CLI)
-
-## Install
-
-```sh
-npm install ng2-alfresco-sites
-```
-
-## Basic usage
-
-In this component is present one single component:
-
-* adf-sites-dropdown
-
-## adf-sites-dropdown
-
-```html
-<adf-sites-dropdown></adf-sites-dropdown>
-``` 
-
-### Properties
-
-There are no input properties at the moment for this component
-
-### Events
-
-| Name | Description |
-| --- | --- |
-| siteSelected | Raised when a site from dropdown is selected |                     
-
-## Build from sources
-
-You can build component from sources with the following commands:
-
-```sh
-npm install
-npm run build
-```
-
-> The `build` task rebuilds all the code, runs tslint, license checks 
-> and other quality check tools before performing unit testing.
-
-## NPM scripts
-
-| Command | Description |
-| --- | --- |
-| npm run build | Build component |
-| npm run test | Run unit tests in the console |
-| npm run test-browser | Run unit tests in the browser
-| npm run coverage | Run unit tests and display code coverage report |
-
-## Demo
-
-Please check the demo folder for a demo project
-
-```sh
-cd demo
-npm install
-npm start
-```
-
-## License
-
-[Apache Version 2.0](https://github.com/Alfresco/alfresco-ng2-components/blob/master/LICENSE)
