--- ng2-components/ng2-alfresco-sites/config/helpers.js
@@ -1,10 +1,0 @@
-var path = require('path');
-
-var _root = path.resolve(__dirname, '..');
-
-function root(args) {
-  args = Array.prototype.slice.call(arguments, 0);
-  return path.join.apply(path, [_root].concat(args));
-}
-
-exports.root = root;
