--- ng2-components/ng2-alfresco-sites/config/webpack.test.js
@@ -1,8 +1,0 @@
-const helpers = require('./helpers');
-const webpackMerge = require('webpack-merge');
-const commonConfig = require('./webpack.common.js');
-
-module.exports = webpackMerge(commonConfig, {
-
-    devtool: 'inline-source-map'
-});
