--- ng2-components/ng2-alfresco-sites/karma-test-shim.js
@@ -1,25 +1,0 @@
-Error.stackTraceLimit = Infinity;
-
-require('core-js/es6');
-require('core-js/es7/reflect');
-
-require('zone.js/dist/zone');
-require('zone.js/dist/long-stack-trace-zone');
-require('zone.js/dist/proxy');
-require('zone.js/dist/sync-test');
-require('zone.js/dist/jasmine-patch');
-require('zone.js/dist/async-test');
-require('zone.js/dist/fake-async-test');
-
-jasmine.DEFAULT_TIMEOUT_INTERVAL = 10000;
-
-var appContext = require.context('./src', true, /\.spec\.ts/);
-appContext.keys().forEach(appContext);
-
-
-var testing = require('@angular/core/testing');
-var browser = require('@angular/platform-browser-dynamic/testing');
-
-testing.TestBed.initTestEnvironment(browser.BrowserDynamicTestingModule, browser.platformBrowserDynamicTesting());
-
-
