--- ng2-components/ng2-alfresco-sites/package.json
@@ -1,107 +1,0 @@
-{
-  "name": "ng2-alfresco-sites",
-  "description": "Alfresco rating component",
-  "version": "1.6.1",
-  "author": "Alfresco Software, Ltd.",
-  "scripts": {
-    "clean": "rimraf dist node_modules typings bundles coverage .npmrc",
-    "clean-lock": "rimraf package-lock.json",
-    "rimraf": "rimraf",
-    "build": "webpack --config webpack.build.js --progress --profile --bail",
-    "test": "karma start karma.conf.js --reporters mocha,coverage --single-run --mode coverage",
-    "test-browser": "karma start karma.conf.js --reporters kjhtml --component",
-    "coverage": "npm run test && wsrv -o -p 9875 ./coverage/report",
-    "prepublishOnly": "npm run build"
-  },
-  "main": "bundles/ng2-alfresco-sites.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Alfresco/alfresco-ng2-components.git"
-  },
-  "bugs": {
-    "url": "https://github.com/Alfresco/alfresco-ng2-components/issues"
-  },
-  "dependencies": {
-    "@angular/animations": "4.2.5",
-    "@angular/common": "4.2.5",
-    "@angular/compiler": "4.2.5",
-    "@angular/core": "4.2.5",
-    "@angular/forms": "4.2.5",
-    "@angular/http": "4.2.5",
-    "@angular/platform-browser": "4.2.5",
-    "@angular/platform-browser-dynamic": "4.2.5",
-    "@angular/router": "4.2.5",
-    "@angular/material": "2.0.0-beta.6",
-    "alfresco-js-api": "~1.6.1",
-    "core-js": "2.4.1",
-    "hammerjs": "2.0.8",
-    "ng2-alfresco-core": "1.6.1",
-    "@ngx-translate/core": "7.0.0",
-    "reflect-metadata": "0.1.10",
-    "rxjs": "5.1.0",
-    "systemjs": "0.19.27",
-    "zone.js": "0.6.26"
-  },
-  "devDependencies": {
-    "@types/hammerjs": "2.0.34",
-    "@types/jasmine": "2.5.35",
-    "@types/node": "6.0.45",
-    "angular2-template-loader": "0.6.2",
-    "autoprefixer": "6.5.4",
-    "codelyzer": "3.1.2",
-    "copy-webpack-plugin": "4.0.1",
-    "css-loader": "0.23.1",
-    "css-to-string-loader": "0.1.2",
-    "cssnano": "3.8.1",
-    "extract-text-webpack-plugin": "2.0.0-rc.3",
-    "file-loader": "0.11.1",
-    "fork-ts-checker-webpack-plugin": "0.2.3",
-    "happypack": "3.0.0",
-    "html-loader": "0.4.4",
-    "html-webpack-plugin": "2.28.0",
-    "istanbul-instrumenter-loader": "0.2.0",
-    "jasmine-ajax": "3.2.0",
-    "jasmine-core": "2.4.1",
-    "karma": "0.13.22",
-    "karma-chrome-launcher": "~1.0.1",
-    "karma-coverage": "1.1.1",
-    "karma-jasmine": "~1.0.2",
-    "karma-jasmine-ajax": "0.1.13",
-    "karma-jasmine-html-reporter": "0.2.0",
-    "karma-mocha-reporter": "2.2.2",
-    "karma-remap-istanbul": "0.6.0",
-    "karma-sourcemap-loader": "0.3.7",
-    "karma-systemjs": "0.16.0",
-    "karma-webpack": "2.0.3",
-    "loader-utils": "1.1.0",
-    "merge-stream": "1.0.1",
-    "node-sass": "4.5.3",
-    "null-loader": "0.1.1",
-    "package-json-merge": "0.0.1",
-    "raw-loader": "0.5.1",
-    "remap-istanbul": "0.6.3",
-    "rimraf": "2.6.1",
-    "run-sequence": "1.2.2",
-    "sass-loader": "6.0.5",
-    "script-loader": "0.7.0",
-    "source-map-loader": "0.1.6",
-    "style-loader": "0.13.1",
-    "systemjs-builder": "0.15.34",
-    "to-string-loader": "1.1.5",
-    "traceur": "0.0.91",
-    "ts-loader": "2.2.1",
-    "ts-node": "1.7.0",
-    "tslint": "4.5.1",
-    "tslint-loader": "3.5.3",
-    "typescript": "2.3.4",
-    "webpack": "2.2.1",
-    "webpack-dev-server": "2.3.0",
-    "webpack-merge": "2.6.1",
-    "wsrv": "0.1.7"
-  },
-  "keywords": [
-    "rating",
-    "alfresco-component"
-  ],
-  "license": "Apache-2.0"
-}
