--- ng2-components/ng2-alfresco-sites/src/i18n/en.json
@@ -1,7 +1,0 @@
-{
-    "DROPDOWN": {
-        "PLACEHOLDER_LABEL": "Site List",
-        "DEFAULT_OPTION": "No Site Chosen"
-    }
-
-}
