--- ng2-components/ng2-alfresco-sites/webpack.coverage.js
@@ -1,1 +1,0 @@
-module.exports = require('./config/webpack.coverage.js');
