--- ng2-components/ng2-alfresco-sites/webpack.test.js
@@ -1,1 +1,0 @@
-module.exports = require('./config/webpack.test.js');
